made forget password form match input behavior as login and sign in
