+++ src/components/resetForm/ResetPasswordForm.js
@@ -76,11 +76,11 @@ const validationSchema = yup.object().shape({
 
 const ResetPasswordForm = withFormik({
   mapPropsToValues: () => ({
-    email: '',
+    confirmPassword: '',
     password: '',
   }),
   handleSubmit: (values, { props, setSubmitting }) => {
-    props.resetPassword(values.password, values.password, props.history);
+    props.resetPassword(values.confirmPassword, values.password, props.history);
     setSubmitting(false);
   },
   validationSchema,
